requirements.txt
django-markdown-deux
django-pagedown
sorl-thumbnail